requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
blake3==1.0.5
blinker==1.9.0
cachetools==6.2.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
Deprecated==1.2.18
dotenv==0.9.9
Flask==3.1.1
Flask-Limiter==3.12
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
gunicorn==20.1.0
httplib2==0.31.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
oauthlib==3.3.1
ordered-set==4.1.0
packaging==25.0
proto-plus==1.26.1
protobuf==6.32.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.5
python-dotenv==1.1.0
requests==2.32.4
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9.1
setuptools==80.9.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.2
gunicorn==20.1.0